Moved GuiTeleplatesList
Made GuiTeleplatesList to inner class of GuiSelectTeleplate.
